requirements.txt
PyQt5
PyQt5_sip
python_magic
Requests
